add scanned list functionality
'show full list' button starts new activity with full list of scanned items

+++ Project/app/src/main/java/ca/unb/mobiledev/appdevproject/InventoryItem.kt
@@ -6,7 +6,7 @@ package ca.unb.mobiledev.appdevproject
 
 data class InventoryItem(
     val name: String,       // eg "square", "circle"
-    val id: Int,       // unique per item of the same type
+    val id: Long,       // unique per item of the same type
     var quantity: Int, // total count of items of this type
     var damaged: Int // count of damaged items of this type
 )
